requirements.txt
python-pptx==0.6.23
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0

